Extract Google Pay availability into new class
COAND-943
